Brings in the numValues widening (7567882955c1, #156445), which
resolves the follow-up this branch had raised, and mirrors it on the
string column so the two stay in step.

numValues is now a long in StringColumnMetadata, written with
writeVLong; numBlocks, value addresses and the block cache widen to
match, with casts only where an array index or the cursor's int
valueCount contract forces one. The plan drops the width item from
Next.

Both consumers go back to plain += now that the accumulator is a long.
Math.addExact guarded the int cap, which no longer exists; keeping it
would have guarded against 2^63 values and left the numeric line
diverged from upstream.

Conflicts were the value-address accessors and the writer's running
counter, all in numeric files that upstream widened while this branch
renamed them. Resolved by taking upstream's types with this branch's
rank and value-address naming.

+++ libs/columnar/src/main/java/org/elasticsearch/columnar/numeric/NumericColumnReader.java
@@ -41,7 +41,7 @@ public final class NumericColumnReader {
     private final NumericBlockEncoder encoder;
     private final long[] blockBuffer;
 
-    private int cachedBlock = -1;
+    private long cachedBlock = -1;
 
     public NumericColumnReader(NumericColumnMetadata meta, IndexInput data) throws IOException {
         this.meta = meta;
@@ -84,20 +84,20 @@ public final class NumericColumnReader {
     }
 
     /** The value address of a document's first value, given its rank. */
-    public int firstValueAddress(int rank) {
-        return valueAddresses == null ? rank : (int) valueAddresses.get(rank);
+    public long firstValueAddress(int rank) {
+        return valueAddresses == null ? rank : valueAddresses.get(rank);
     }
 
     /** The number of values a document has, given its rank. */
-    public int valueCount(int rank) {
-        return valueAddresses == null ? 1 : (int) (valueAddresses.get(rank + 1) - valueAddresses.get(rank));
+    public long valueCount(int rank) {
+        return valueAddresses == null ? 1 : valueAddresses.get(rank + 1) - valueAddresses.get(rank);
     }
 
     /** The value at {@code valueAddress} in {@code [0, numValues)}. */
-    public long valueAt(int valueAddress) throws IOException {
-        int block = valueAddress / meta.blockSize();
+    public long valueAt(long valueAddress) throws IOException {
+        long block = valueAddress / meta.blockSize();
         ensureBlock(block);
-        return blockBuffer[valueAddress - block * meta.blockSize()];
+        return blockBuffer[(int) (valueAddress - block * meta.blockSize())];
     }
 
     /** Values per encoding block. */
@@ -106,7 +106,7 @@ public final class NumericColumnReader {
     }
 
     /** Total number of values across all documents. */
-    public int numValues() {
+    public long numValues() {
         return meta.numValues();
     }
 
@@ -114,12 +114,12 @@ public final class NumericColumnReader {
      * Decodes the block at {@code blockIndex} (single-block cache) and returns the shared buffer, valid
      * until the next call that touches a different block.
      */
-    public long[] block(int blockIndex) throws IOException {
+    public long[] block(long blockIndex) throws IOException {
         ensureBlock(blockIndex);
         return blockBuffer;
     }
 
-    private void ensureBlock(int block) throws IOException {
+    private void ensureBlock(long block) throws IOException {
         if (block == cachedBlock) {
             return;
         }
@@ -129,7 +129,7 @@ public final class NumericColumnReader {
         int length = (int) (blockEnd - blockStart);
         DataInput blockData = blockBytesCodec.read(data, length);
         // Full blocks hold blockSize values; the last block holds the remainder.
-        int valueCount = Math.min(meta.blockSize(), meta.numValues() - block * meta.blockSize());
+        int valueCount = (int) Math.min(meta.blockSize(), meta.numValues() - block * meta.blockSize());
         encoder.decode(blockData, valueCount, blockBuffer);
         cachedBlock = block;
     }
